+++ admin/controllers/ListingBaseController.php
@@ -39,7 +39,7 @@ class ListingBaseController extends AdminBaseController
 
         $recordCountSrch->doNotLimitRecords();
         if ($isGroupSearch == false) {
-            $recordCountSrch->addFld('count(*) as totalRecords');
+            $recordCountSrch->addFld('count(1) as totalRecords');
             $recordCountSrch->doNotCalculateRecords();
             $results = FatApp::getDb()->fetch($recordCountSrch->getResultSet());
             $defaultRecordCount = !empty($results['totalRecords']) ? $results['totalRecords'] : 0;