+++ dashboard/controllers/DashboardBaseController.php
@@ -830,7 +830,7 @@ class DashboardBaseController extends FatController
 
         $recordCountSrch->doNotLimitRecords();
         if ($isGroupSearch == false) {
-            $recordCountSrch->addFld('count(*) as totalRecords');
+            $recordCountSrch->addFld('count(1) as totalRecords');
             $recordCountSrch->doNotCalculateRecords();
             $results = FatApp::getDb()->fetch($recordCountSrch->getResultSet());
             $defaultRecordCount = !empty($results['totalRecords']) ? $results['totalRecords'] : 0;